refactor: break out click to seperate file and move focus stuff to a11y
