+++ src/table/components/TableTotals.tsx
@@ -2,7 +2,7 @@ import React, { memo, useMemo } from 'react';
 import { useContextSelector, TableContext } from '../context';
 import { getTotalsCellStyle } from '../utils/styling-utils';
 import { handleTotalKeyDown } from '../utils/handle-key-press';
-import { removeAndFocus } from '../utils/handle-accessibility';
+import { removeTabAndFocusCell } from '../utils/accessibility-utils';
 import { StyledHeadRow, StyledTotalsCell } from '../styles';
 import { TableTotalsProps } from '../types';
 
@@ -30,7 +30,7 @@ function TableTotals({ rootElement, tableData, theme, layout, keyboard, selectio
               handleTotalKeyDown(e, rootElement, cellCoord, setFocusedCellCoord, selectionsAPI.isModal());
             }}
             onMouseDown={() => {
-              removeAndFocus(cellCoord, rootElement, setFocusedCellCoord, keyboard);
+              removeTabAndFocusCell(cellCoord, rootElement, setFocusedCellCoord, keyboard);
             }}
           >
             {column.totalInfo}
