+++ src/table/utils/accessibility-utils.ts
@@ -1,6 +1,5 @@
-import { MouseEvent } from 'react';
 import { stardust } from '@nebula.js/stardust';
-import { Announce, Cell, TotalsPosition } from '../../types';
+import { Announce } from '../../types';
 import { CellFocusProps, HandleResetFocusProps } from '../types';
 
 export const getCellElement = (rootElement: HTMLElement, cellCoord: [number, number]) =>
@@ -11,6 +10,9 @@ export const getCellElement = (rootElement: HTMLElement, cellCoord: [number, num
 export const findCellWithTabStop = (rootElement: HTMLElement) =>
   rootElement.querySelector("td[tabindex='0'], th[tabindex='0']") as HTMLTableCellElement;
 
+/**
+ * Removes/adds tab stop and sometimes focuses or blurs the cell, depending on focusTupe
+ */
 export const updateFocus = ({ focusType, cell }: CellFocusProps) => {
   if (!cell) return;
 
@@ -34,7 +36,82 @@ export const updateFocus = ({ focusType, cell }: CellFocusProps) => {
   }
 };
 
-export const removeAndFocus = (
+/**
+ * Calculates the next cell to focus
+ */
+export const getNextCellCoord = (
+  evt: React.KeyboardEvent,
+  rootElement: HTMLElement,
+  cellCoord: [number, number],
+  allowedRows: {
+    top: number;
+    bottom: number;
+  } = { top: 0, bottom: 0 }
+): [number, number] => {
+  const rowCount = rootElement.getElementsByClassName('sn-table-row').length;
+  const columnCount = rootElement.getElementsByClassName('sn-table-head-cell').length;
+  let [nextRow, nextCol] = cellCoord;
+
+  switch (evt.key) {
+    case 'ArrowDown':
+      nextRow < rowCount - 1 - allowedRows.bottom && nextRow++;
+      break;
+    case 'ArrowUp':
+      nextRow > allowedRows.top && nextRow--;
+      break;
+    case 'ArrowRight':
+      // allowedRows.top greater than 0 means we are in selection mode
+      if (allowedRows.top > 0) break;
+      if (nextCol < columnCount - 1) {
+        nextCol++;
+      } else if (nextRow < rowCount - 1) {
+        nextRow++;
+        nextCol = 0;
+      }
+      break;
+    case 'ArrowLeft':
+      // allowedRows.top greater than 0 means we are in selection mode
+      if (allowedRows.top > 0) break;
+      if (nextCol > 0) {
+        nextCol--;
+      } else if (nextRow > 0) {
+        nextRow--;
+        nextCol = columnCount - 1;
+      }
+      break;
+    default:
+      break;
+  }
+
+  return [nextRow, nextCol];
+};
+
+/**
+ * Resets and adds new focus to a table cell based which key is pressed
+ */
+export const moveFocus = (
+  evt: React.KeyboardEvent,
+  rootElement: HTMLElement,
+  cellCoord: [number, number],
+  setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
+  allowedRows?: {
+    top: number;
+    bottom: number;
+  }
+) => {
+  updateFocus({ focusType: 'removeTab', cell: evt.target as HTMLTableCellElement });
+  const nextCellCoord = getNextCellCoord(evt, rootElement, cellCoord, allowedRows);
+  const nextCell = getCellElement(rootElement, nextCellCoord);
+  updateFocus({ focusType: 'focus', cell: nextCell });
+  setFocusedCellCoord(nextCellCoord);
+
+  return nextCell;
+};
+
+/**
+ * Resets and adds new focus to the cell at position newCoord
+ */
+export const removeTabAndFocusCell = (
   newCoord: [number, number],
   rootElement: HTMLElement,
   setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
@@ -45,33 +122,10 @@ export const removeAndFocus = (
   keyboard.enabled && keyboard.focus?.();
 };
 
-export const handleClickToFocusBody = (
-  cell: Cell,
-  rootElement: HTMLElement,
-  setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
-  keyboard: stardust.Keyboard,
-  totalsPosition: TotalsPosition
-) => {
-  const { rawRowIdx, rawColIdx } = cell;
-  const adjustedRowIdx = totalsPosition === 'top' ? rawRowIdx + 2 : rawRowIdx + 1;
-  removeAndFocus([adjustedRowIdx, rawColIdx], rootElement, setFocusedCellCoord, keyboard);
-};
-
-export const handleClickToFocusHead = (
-  columnIndex: number,
-  rootElement: HTMLElement,
-  setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
-  keyboard: stardust.Keyboard
-) => {
-  removeAndFocus([0, columnIndex], rootElement, setFocusedCellCoord, keyboard);
-};
-
-export const handleMouseDownLabelToFocusHeadCell = (evt: MouseEvent, rootElement: HTMLElement, columnIndex: number) => {
-  evt.preventDefault();
-  updateFocus({ focusType: 'focus', cell: getCellElement(rootElement, [0, columnIndex]) });
-};
-
-export const handleResetFocus = ({
+/**
+ * Resets the focus when the data size or page is change, only adds a tab stop when shouldRefocus.current is false
+ */
+export const resetFocus = ({
   focusedCellCoord,
   rootElement,
   shouldRefocus,
@@ -85,6 +139,7 @@ export const handleResetFocus = ({
   // If you have selections ongoing, you want to stay on the same column
   const selectionCellCoord: [number, number] = [totalsPosition === 'top' ? 2 : 1, focusedCellCoord[1]];
   const cellCoord: [number, number] = isSelectionMode ? selectionCellCoord : [0, 0];
+
   if (!keyboard.enabled || keyboard.active) {
     // Only run this if updates come from inside table
     const focusType = shouldRefocus.current ? 'focus' : 'addTab';
@@ -102,9 +157,13 @@ export const handleResetFocus = ({
       });
     }
   }
+
   setFocusedCellCoord(cellCoord);
 };
 
+/**
+ * On focus is no longer in the table, resets the announcer and calls keyboard.blur
+ */
 export const handleFocusoutEvent = (
   evt: FocusEvent,
   shouldRefocus: React.MutableRefObject<boolean>,
@@ -121,6 +180,9 @@ export const handleFocusoutEvent = (
   }
 };
 
+/**
+ * Sets focus to button in the selection toolbar, handles both client and nebula toolbar
+ */
 export const focusSelectionToolbar = (element: HTMLElement, keyboard: stardust.Keyboard, last: boolean) => {
   const clientConfirmButton = element
     .closest('.qv-object-wrapper')
@@ -133,6 +195,9 @@ export const focusSelectionToolbar = (element: HTMLElement, keyboard: stardust.K
   keyboard.focusSelection?.(last);
 };
 
+/**
+ * Updates the announcer with current cell selection state
+ */
 export const announceSelectionState = (
   announce: Announce,
   nextCell: HTMLTableCellElement,
