+++ src/table/utils/handle-click.ts
@@ -1,0 +1,30 @@
+import { MouseEvent } from 'react';
+import { stardust } from '@nebula.js/stardust';
+import { Cell, TotalsPosition } from '../../types';
+import { removeTabAndFocusCell, updateFocus, getCellElement } from './accessibility-utils';
+
+export const handleClickToFocusBody = (
+  cell: Cell,
+  rootElement: HTMLElement,
+  setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
+  keyboard: stardust.Keyboard,
+  totalsPosition: TotalsPosition
+) => {
+  const { rawRowIdx, rawColIdx } = cell;
+  const adjustedRowIdx = totalsPosition === 'top' ? rawRowIdx + 2 : rawRowIdx + 1;
+  removeTabAndFocusCell([adjustedRowIdx, rawColIdx], rootElement, setFocusedCellCoord, keyboard);
+};
+
+export const handleClickToFocusHead = (
+  columnIndex: number,
+  rootElement: HTMLElement,
+  setFocusedCellCoord: React.Dispatch<React.SetStateAction<[number, number]>>,
+  keyboard: stardust.Keyboard
+) => {
+  removeTabAndFocusCell([0, columnIndex], rootElement, setFocusedCellCoord, keyboard);
+};
+
+export const handleMouseDownLabelToFocusHeadCell = (evt: MouseEvent, rootElement: HTMLElement, columnIndex: number) => {
+  evt.preventDefault();
+  updateFocus({ focusType: 'focus', cell: getCellElement(rootElement, [0, columnIndex]) });
+};
